requirements.txt
matplotlib
numpy
pandas
scipy
pillow